requirements.txt
mutils
opencv-python
numpy
request
keyboard
pypiwin32
pygame
Pillow
imutils
pyautogui
